split out open authoring logic from index view
